Components naming update
